Setup do projeto e criação de testes
