+++ src/main/java/br/calebe/ticketmachine/exception/SaldoInsuficienteException.java
@@ -1,0 +1,12 @@
+package br.calebe.ticketmachine.exception;
+
+/**
+ *
+ * @author Calebe de Paula Bianchini
+ */
+public class SaldoInsuficienteException extends Exception {
+
+    public SaldoInsuficienteException(String string) {
+        System.out.println(string);
+    }
+}
